Refactoring. Move inspectors to separate package
