+++ winter_openapi/inspectors/enum_inspector.py
@@ -1,7 +1,7 @@
 from enum import Enum
 from typing import Type
 
-from .type_inspection import inspect_enum
+from winter_openapi.type_inspection import inspect_enum
 
 
 def inspect_enum_class(enum_class: Type[Enum]) -> dict:
